Dynamic YAML/regex testing expanded to all Windows YAML files.
