+++ tests/process_scripts/regex/windows/test_windows_logging.py
@@ -1,0 +1,23 @@
+"""
+Dynamically generated tests for audit-windows-logging.yaml regex patterns.
+
+This module automatically creates unit tests for each regex pattern defined
+in the audit-windows-logging.yaml configuration file, testing against real
+Windows system data.
+"""
+
+import pytest
+
+from tests.process_scripts.regex.windows.test_all_windows_dynamic import (
+    DynamicWindowsYamlTestMixin,
+)
+
+
+class TestWindowsLoggingPatterns(DynamicWindowsYamlTestMixin):
+    """Dynamic tests for audit-windows-logging.yaml regex patterns."""
+
+    YAML_FILENAME = "audit-windows-logging.yaml"
+
+
+if __name__ == "__main__":
+    pytest.main([__file__, "-v"])
